+++ Backbuffer.cpp
@@ -14,11 +14,11 @@ Backbuffer::~Backbuffer()
 
 void Backbuffer::flip() const noexcept
 {
-	const Vector2 pos{ 0,0 };
-	const Rectangle source = { 0, 0, _backbuffer.texture.width, -_backbuffer.texture.height };
+	const Vector2 pos{ 0.0f, 0.0f };
+	const Rectangle source = { 0.0f, 0.0f, static_cast<float>(_backbuffer.texture.width), -static_cast<float>(_backbuffer.texture.height) };
 
 	BeginDrawing();
-	const Rectangle dest = { 0, 0, _backbuffer.texture.width * _scale, _backbuffer.texture.height * _scale};
+	const Rectangle dest = { 0.0f, 0.0f, static_cast<float>(_backbuffer.texture.width) * _scale, static_cast<float>(_backbuffer.texture.height) * _scale };
 	DrawTexturePro(_backbuffer.texture, source, dest, pos, 0.0F, WHITE);
 	EndDrawing();
 }
